+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component } from '@angular/core';
+import { Component, TemplateRef, ViewChild } from '@angular/core';
 import {
   DefaultLayoutComponent
 } from '@components/default-layout/default-layout.component';
@@ -15,5 +15,7 @@ import { ThemeService } from '@services/theme.service';
   templateUrl: './app.component.html'
 })
 export class AppComponent {
+  @ViewChild('mainElement') mainElement!: TemplateRef<any>;
+
   constructor(public theme: ThemeService) {}
 }